[#116] Add coverage for integration tests
