+++ .github/workflows/integration_test.yml
@@ -38,4 +38,4 @@ jobs:
       - name: Run integration tests
         run: |
           cd test
-          go test ./...
+          go test -v ./... -cover -coverpkg=../... -coverprofile cover.out  && go tool cover -func cover.out
